move saveValidatorChains
